maybe

+++ Entrega2/java_doctic_project/src/main/java/com/platform/doctic_project/Controller/CategoriaController.java
@@ -1,12 +1,17 @@
 package com.platform.doctic_project.Controller;
 
-import com.platform.doctic_project.Model.Categoria;
-import com.platform.doctic_project.Service.CategoriaServiceImp;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import com.platform.doctic_project.Model.Categoria;
+import com.platform.doctic_project.Service.CategoriaServiceImp;
 
 @RestController
 @RequestMapping("/api/categories")
@@ -22,7 +27,7 @@ public class CategoriaController {
     }
 
     @PostMapping("/assign")
-    public ResponseEntity<String> assignCategoryToDocument(@RequestParam Long documentId, @RequestParam Long categoryId) {
+    public ResponseEntity<String> assignCategoryToDocument(@RequestParam Integer documentId, @RequestParam Integer categoryId) {
         categoriaServiceImp.assignCategoryToDocument(documentId, categoryId);
         return ResponseEntity.ok("Categoría asignada al documento con éxito.");
     }
