+++ Entrega2/java_doctic_project/src/main/java/com/platform/doctic_project/Controller/RecomendacionController.java
@@ -1,13 +1,17 @@
 package com.platform.doctic_project.Controller;
 
-import com.platform.doctic_project.Model.Documento;
-import com.platform.doctic_project.Service.RecomendacionServiceImp;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import com.platform.doctic_project.Model.Documento;
+import com.platform.doctic_project.Service.RecomendacionServiceImp;
 
 @RestController
 @RequestMapping("/api/v1/recomendaciones") // Endpoints base para recomendaciones
@@ -18,7 +22,7 @@ public class RecomendacionController {
 
     // Generar recomendaciones de documentos para un usuario
     @GetMapping("/generar/{userId}")
-    public ResponseEntity<?> generateDocumentRecommendations(@PathVariable Long userId) {
+    public ResponseEntity<?> generateDocumentRecommendations(@PathVariable Integer userId) {
         try {
             List<Documento> recomendaciones = recomendacionService.generateDocumentRecommendations(userId);
             return ResponseEntity.ok(recomendaciones);
