+++ Entrega2/java_doctic_project/src/main/java/com/platform/doctic_project/Controller/VisualizacionController.java
@@ -1,12 +1,18 @@
 package com.platform.doctic_project.Controller;
 
-import com.platform.doctic_project.Service.IVisualizacionService;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import com.platform.doctic_project.Service.IVisualizacionService;
 
 @RestController
 @RequestMapping("/api/v1/visualizaciones") // Endpoint base para visualizaciones
@@ -17,7 +23,7 @@ public class VisualizacionController {
 
     // Registrar la visualización de un documento
     @PostMapping("/registrar")
-    public ResponseEntity<String> recordDocumentView(@RequestParam Long userId, @RequestParam Long documentId) {
+    public ResponseEntity<String> recordDocumentView(@RequestParam Integer userId, @RequestParam Integer documentId) {
         try {
             visualizacionService.recordDocumentView(userId, documentId);
             return new ResponseEntity<>("Visualización del documento registrada con éxito.", HttpStatus.OK);
@@ -28,9 +34,9 @@ public class VisualizacionController {
 
     // Obtener el historial de visualizaciones de un usuario
     @GetMapping("/usuario/{userId}")
-    public ResponseEntity<?> getUserViewHistory(@PathVariable Long userId) {
+    public ResponseEntity<?> getUserViewHistory(@PathVariable Integer userId) {
         try {
-            List<Long> viewHistory = visualizacionService.getUserViewHistory(userId);
+            List<Integer> viewHistory = visualizacionService.getUserViewHistory(userId);
             return ResponseEntity.ok(viewHistory);
         } catch (Exception e) {
             return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
