+++ Entrega2/java_doctic_project/src/main/java/com/platform/doctic_project/Service/IVisualizacionService.java
@@ -3,6 +3,6 @@ package com.platform.doctic_project.Service;
 import java.util.List;
 
 public interface IVisualizacionService {
-    void recordDocumentView(Long userId, Long documentId);
-    List<Long> getUserViewHistory(Long userId);
+    void recordDocumentView(Integer userId, Integer documentId);
+    List<Integer> getUserViewHistory(Integer userId);
 }
